Reapply "Extraceted the PlayerDeathListener and EnderdragonDeathListener from the ChallengeEndListener and added methods to Challenge which get active challenges as a String and as a List of challenges"
This reverts commit e988373915876568ddec916a7e840e98525a6415.
